// juridico_angular/src/app/models/pedido.ts
export interface Pedido {
   codigo: number;
   pedido: string;
   vlrPedido: number;
   primeiraInst: string;
   primeiraPrev: string;
   primeiraDecisao: string;
   segundaInst: string;
   segundaPrev: string;
   segundaDecisao: string;
   terceiraInst: string;
   terceiraPrev: string;
   terceiraDecisao: string;
}